Dockerfile: Pin Node
This patch pins the Node version in `docker/Dockerfile`.

Additionally, this patch explicitly references the image registry that
was implicitly referenced before.

Since `node:23` is currently an alias of `node:latest`, this patch
should effectively be a no-op.

Other than regular code hygiene, this patch may or may not turn out
useful down the road for caching builds.